Changed the train code to use a mapping from encoded coordinates to track pieces in order to improve efficiency and enable use of AStarPathUtil to find paths for the Engineer card.

+++ src/java/com/threerings/bang/game/data/piece/Piece.java
@@ -59,7 +59,7 @@ public abstract class Piece extends SimpleStreamableObject
     /**
      * Combines the supplied x and y coordintes into a single integer.
      */
-    public static int coord (short x, short y)
+    public static int coord (int x, int y)
     {
         return (x << 16) | y;
     }
